Added the build tasks for distribution
:) Yay! Following TreeHouse’s workshop on using useref I’ve added the
tasks needed for deployment: ‘clean’, ‘html’ and ‘build’ (see the
comments in the gulp file to learn more).

I’d like to keep the ‘build’ task as is (so not added as a dependency
of the default task, which is reserved for development), so I have a
clear action (‘gulp build’) whenever I’m ready for deployment.