refactored styles for .section-content and .card
Be wary of flex box support for Safari. When flex-direction: column is
declared the containing element doesn’t grow with it’s content.

Also, min-height just becomes height, neglecting the height of the
actual content.

Super weird and I still have to figure out what exactly is going on
here.